Making the repository public
Much work is still needed before I decide it's a good idea to add more features.

The features already present will need to be reworked in order to get the code to a more maintainable level before more stuff is added. I'd hate to have to start over again because the code is just unusable.

The next push will be mostly code clean up and maybe the rewriting of some parts.