Mother of all refactors
Fell into the OOP refactor rabbit hole.

Most of the stuff I changed wasn't even necessary, but I guess the code is more concise and readable now.

Never again do I want to do this, but I have a feeling it will happen; specially with the tests.

I need to find a better way to test. To only test the important stuff.